@@ -6,13 +6,13 @@
 
     <groupId>net.toiletmc</groupId>
     <artifactId>LagManager</artifactId>
-    <version>1.1.4</version>
+    <version>1.2</version>
     <packaging>jar</packaging>
 
     <name>LagManager</name>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
     <url>https://toiletmc.net</url>
@@ -73,12 +73,6 @@
     </repositories>
 
     <dependencies>
-        <dependency>
-            <groupId>org.spigotmc</groupId>
-            <artifactId>plugin-annotations</artifactId>
-            <version>1.2.3-SNAPSHOT</version>
-            <scope>provided</scope>
-        </dependency>
         <dependency>
             <groupId>io.papermc.paper</groupId>
             <artifactId>paper-api</artifactId>
